% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_batch_describe_simulation_job}
\alias{robomaker_batch_describe_simulation_job}
\title{Describes one or more simulation jobs}
\usage{
robomaker_batch_describe_simulation_job(jobs)
}
\arguments{
\item{jobs}{[required] A list of Amazon Resource Names (ARNs) of simulation jobs to describe.}
}
\description{
Describes one or more simulation jobs.
}
\section{Request syntax}{
\preformatted{robomaker$batch_describe_simulation_job(
  jobs = list(
    "string"
  )
)
}
}

\keyword{internal}
